TIP-856: Move Pim\Versionning class to Akeneo\Tool
