+++ src/Pim/Bundle/EnrichBundle/spec/Provider/StructureVersion/StructureVersionProviderSpec.php
@@ -2,7 +2,7 @@
 
 namespace spec\Pim\Bundle\EnrichBundle\Provider\StructureVersion;
 
-use Akeneo\Component\Versioning\Model\Version;
+use Akeneo\Tool\Component\Versioning\Model\Version;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\VersioningBundle\Repository\VersionRepositoryInterface;
