+++ src/Pim/Bundle/VersioningBundle/Command/RefreshCommand.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\VersioningBundle\Command;
 
-use Akeneo\Component\Versioning\Model\Version;
+use Akeneo\Tool\Component\Versioning\Model\Version;
 use Akeneo\Tool\Component\StorageUtils\Detacher\ObjectDetacherInterface;
 use Doctrine\Common\Persistence\ObjectManager;
 use Monolog\Handler\StreamHandler;
