+++ src/Pim/Bundle/VersioningBundle/Manager/VersionManager.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\VersioningBundle\Manager;
 
-use Akeneo\Component\Versioning\Model\Version;
+use Akeneo\Tool\Component\Versioning\Model\Version;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Persistence\ObjectManager;
 use Doctrine\Common\Util\ClassUtils;
