+++ src/Pim/Bundle/VersioningBundle/Purger/VersionPurgerAdvisorInterface.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\VersioningBundle\Purger;
 
-use Akeneo\Component\Versioning\Model\VersionInterface;
+use Akeneo\Tool\Component\Versioning\Model\VersionInterface;
 
 /**
  * Checks if a version should be purged
