+++ src/Pim/Bundle/VersioningBundle/Resources/config/entities.yml
@@ -1,3 +1,3 @@
 parameters:
-    pim_versioning.entity.version.class: Akeneo\Component\Versioning\Model\Version
+    pim_versioning.entity.version.class: Akeneo\Tool\Component\Versioning\Model\Version
     pim_versioning.versionable_entities: ~
