+++ src/Pim/Bundle/VersioningBundle/UpdateGuesser/VersionableUpdateGuesser.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\VersioningBundle\UpdateGuesser;
 
-use Akeneo\Component\Versioning\Model\VersionableInterface;
+use Akeneo\Tool\Component\Versioning\Model\VersionableInterface;
 use Doctrine\Common\Util\ClassUtils;
 use Doctrine\ORM\EntityManager;
